Sort the columns of taxa_table (#22)
For some reason, the `id` column is not necessarily the first one in the
data table for taxa when exporting to igraph. This can cause duplicated
vertex errors.

It would be oh so simple if igraph stopped thinking about positions in
the data.frames and LOOKED AT THE DAMN NAMES.